+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # MAuth-Client History
 
-## v4.0.2
+## v4.0.4
 - Restore original behavior in the proxy of forwarding of headers that begin with HTTP_ (except for HTTP_HOST) but removing the HTTP_.
+
+## v4.0.3
+- Updated signature to decode number sign (#) in requests
+
+## v4.0.2
+- Store the config data to not load the config file multiple times
 
 ## v4.0.1
 - Open source and publish this gem on rubygems.org, no functionality changes